Complete exercise 4.21

+++ controllers/blogs.js
@@ -42,8 +42,18 @@ blogsRouter.post('/', async (req, res, next) => {
 
 blogsRouter.delete('/:id', async (req, res) => {
   try {
-    await Blog.findByIdAndRemove(req.params.id)
-    res.status(204).end()
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    if (!req.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(req.params.id)
+
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(req.params.id)
+      res.status(204).end()
+    }
   } catch {
     res.status(400).json({ 'error': 'blog not found' })
   }
